Revert "platform: Payment observer implementation (#179)"
This reverts commit ae47f7aa901edf0fe2715b9fb4bc1da34215d3e6.
