+++ data-spring-jdbc/src/main/java/org/stellar/anchor/server/data/JdbcSep31TransactionStore.java
@@ -3,7 +3,6 @@ package org.stellar.anchor.server.data;
 import java.time.Instant;
 import java.util.Collection;
 import java.util.List;
-import java.util.Optional;
 import lombok.NonNull;
 import org.stellar.anchor.exception.SepException;
 import org.stellar.anchor.model.Sep31Transaction;
@@ -57,10 +56,5 @@ public class JdbcSep31TransactionStore implements Sep31TransactionStore {
     }
 
     return transactionRepo.save((JdbcSep31Transaction) transaction);
-  }
-
-  public Sep31Transaction findByStellarAccountId(String accountId) {
-    Optional<JdbcSep31Transaction> optTxn = transactionRepo.findByStellarAccountId(accountId);
-    return optTxn.orElse(null);
   }
 }
