+++ platform/src/main/java/org/stellar/anchor/platform/ConfigManagementConfig.java
@@ -4,7 +4,15 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.stellar.anchor.config.*;
+import org.stellar.anchor.config.AppConfig;
+import org.stellar.anchor.config.Sep10Config;
+import org.stellar.anchor.config.Sep1Config;
+import org.stellar.anchor.config.Sep38Config;
 import org.stellar.anchor.server.config.*;
+import org.stellar.anchor.server.config.PropertyAppConfig;
+import org.stellar.anchor.server.config.PropertySep10Config;
+import org.stellar.anchor.server.config.PropertySep1Config;
+import org.stellar.anchor.server.config.PropertySep38Config;
 
 @Configuration
 public class ConfigManagementConfig {
