+++ platform/src/main/java/org/stellar/anchor/platform/RestFeeIntegration.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.platform.callback;
+package org.stellar.anchor.platform;
 
 import static okhttp3.HttpUrl.get;
 import static org.stellar.anchor.platform.PlatformIntegrationHelper.*;
@@ -8,35 +8,38 @@ import java.lang.reflect.Type;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Map;
-import okhttp3.*;
-import okhttp3.HttpUrl.Builder;
+import okhttp3.HttpUrl;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
 import org.springframework.http.HttpStatus;
-import org.stellar.anchor.exception.*;
-import org.stellar.anchor.integration.rate.GetRateRequest;
-import org.stellar.anchor.integration.rate.GetRateResponse;
-import org.stellar.anchor.integration.rate.RateIntegration;
+import org.stellar.anchor.exception.AnchorException;
+import org.stellar.anchor.exception.ServerErrorException;
+import org.stellar.anchor.integration.fee.FeeIntegration;
+import org.stellar.platform.apis.callbacks.requests.GetFeeRequest;
+import org.stellar.platform.apis.callbacks.responses.GetFeeResponse;
 import shadow.com.google.common.reflect.TypeToken;
 
-public class RestRateIntegration implements RateIntegration {
-  private final String anchorEndpoint;
+public class RestFeeIntegration implements FeeIntegration {
+  private final String feeIntegrationEndPoint;
   private final OkHttpClient httpClient;
   private final Gson gson;
 
-  public RestRateIntegration(String anchorEndpoint, OkHttpClient httpClient, Gson gson) {
+  public RestFeeIntegration(String feeIntegrationEndPoint, OkHttpClient okHttpClient, Gson gson) {
     try {
-      new URI(anchorEndpoint);
+      new URI(feeIntegrationEndPoint);
     } catch (URISyntaxException e) {
       throw new IllegalArgumentException("invalid 'baseUri'");
     }
 
-    this.anchorEndpoint = anchorEndpoint;
-    this.httpClient = httpClient;
+    this.feeIntegrationEndPoint = feeIntegrationEndPoint;
+    this.httpClient = okHttpClient;
     this.gson = gson;
   }
 
   @Override
-  public GetRateResponse getRate(GetRateRequest request) throws AnchorException {
-    Builder urlBuilder = get(anchorEndpoint).newBuilder().addPathSegment("rate");
+  public GetFeeResponse getFee(GetFeeRequest request) throws AnchorException {
+    HttpUrl.Builder urlBuilder = get(feeIntegrationEndPoint).newBuilder().addPathSegment("fee");
     Type type = new TypeToken<Map<String, ?>>() {}.getType();
     Map<String, String> paramsMap = gson.fromJson(gson.toJson(request), type);
     paramsMap.forEach(
@@ -56,22 +59,13 @@ public class RestRateIntegration implements RateIntegration {
       throw httpError(responseContent, response.code(), gson);
     }
 
-    GetRateResponse getRateResponse;
+    GetFeeResponse feeResponse;
     try {
-      getRateResponse = gson.fromJson(responseContent, GetRateResponse.class);
+      feeResponse = gson.fromJson(responseContent, GetFeeResponse.class);
     } catch (Exception e) { // cannot read body from response
       throw new ServerErrorException("internal server error", e);
     }
 
-    GetRateResponse.Rate rate = getRateResponse.getRate();
-    if (rate == null || rate.getPrice() == null) {
-      throw new ServerErrorException("internal server error");
-    }
-    if (request.getType() == GetRateRequest.Type.FIRM) {
-      if (rate.getId() == null || rate.getExpiresAt() == null) {
-        throw new ServerErrorException("internal server error");
-      }
-    }
-    return getRateResponse;
+    return feeResponse;
   }
 }
