--- platform/src/main/java/org/stellar/anchor/platform/paymentobserver/PageTokenStore.java
@@ -1,7 +1,0 @@
-package org.stellar.anchor.platform.paymentobserver;
-
-public interface PageTokenStore {
-  void save(String account, String token);
-
-  String load(String account);
-}
