--- platform/src/main/java/org/stellar/anchor/platform/paymentobserver/PaymentListener.java
@@ -1,9 +1,0 @@
-package org.stellar.anchor.platform.paymentobserver;
-
-import org.stellar.sdk.responses.operations.PaymentOperationResponse;
-
-public interface PaymentListener {
-  void onReceived(PaymentOperationResponse payment);
-
-  void onSent(PaymentOperationResponse payment);
-}
