+++ platform/src/main/resources/assets-test.json
@@ -4,7 +4,6 @@
             "schema": "stellar",
             "code": "USDC",
             "issuer": "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
-            "distribution_account": "GBJDTHT4562X2H37JMOE6IUTZZSDU6RYGYUNFYCHVFG3J4MYJIMU33HK",
             "significant_decimals": 2,
             "deposit" : {
                 "enabled": true,
